+++ app/javascript/shared/components/MinimalProfilePreviewCard.jsx
